+++ client-app/src/desktop/common/grid/options/AgGridOptions.js
@@ -2,8 +2,7 @@ import {hoistCmp, uses, XH} from '@xh/hoist/core';
 import {select, switchInput} from '@xh/hoist/desktop/cmp/input';
 import {AgGridModel} from '@xh/hoist/cmp/ag-grid';
 import {fragment, label, vspacer} from '@xh/hoist/cmp/layout';
-import {button} from '@xh/hoist/desktop/cmp/button';
-import {Icon} from '@xh/hoist/icon';
+
 
 export const agGridOptions = hoistCmp.factory({
     model: uses(AgGridModel),
@@ -58,26 +57,6 @@ export const agGridOptions = hoistCmp.factory({
                     bind: 'showCellFocus',
                     label: 'Cell focus',
                     labelAlign: 'left'
-                }),
-                vspacer(10),
-                button({
-                    text: 'Save grid state',
-                    icon: Icon.save(),
-                    minimal: false,
-                    onClick: () => {
-                        const state = model.getState();
-                        console.log(state);
-                        XH.localStorageService.set('agGridWrapperState', state);
-                    }
-                }),
-                button({
-                    text: 'Load grid state',
-                    icon: Icon.grid(),
-                    minimal: false,
-                    onClick: () => {
-                        const state = XH.localStorageService.get('agGridWrapperState');
-                        model.setState(state);
-                    }
                 })
             ]
         });
