+++ client-app/src/desktop/common/grid/options/AgGridOptionsPanel.js
@@ -1,9 +1,10 @@
-import {div} from '@xh/hoist/cmp/layout';
-import {hoistCmp, uses} from '@xh/hoist/core';
+import {div, vspacer} from '@xh/hoist/cmp/layout';
+import {hoistCmp, uses, XH} from '@xh/hoist/core';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {Icon} from '@xh/hoist/icon';
-import {agGridOptions} from './AgGridOptions';
 import {AgGridModel} from '@xh/hoist/cmp/ag-grid';
+import {button} from '@xh/hoist/desktop/cmp/button';
+import {agGridOptions} from './AgGridOptions';
 
 export const agGridOptionsPanel = hoistCmp.factory({
     model: uses(AgGridModel),
@@ -17,7 +18,29 @@ export const agGridOptionsPanel = hoistCmp.factory({
             model: {side: 'right', defaultSize: 250, resizable: false},
             item: div({
                 className: 'tbox-display-opts__inner',
-                item: agGridOptions()
+                item: [
+                    agGridOptions(),
+                    vspacer(10),
+                    button({
+                        text: 'Save grid state',
+                        icon: Icon.save(),
+                        minimal: false,
+                        onClick: () => {
+                            const state = model.getState();
+                            console.log(state);
+                            XH.localStorageService.set('agGridWrapperState', state);
+                        }
+                    }),
+                    button({
+                        text: 'Load grid state',
+                        icon: Icon.grid(),
+                        minimal: false,
+                        onClick: () => {
+                            const state = XH.localStorageService.get('agGridWrapperState');
+                            model.setState(state);
+                        }
+                    })
+                ]
             })
         });
     }
